Move BankSearch to bank plugin
It contains no bank tab specific code and is used by both the bank and banktags plugins
